organized login and register frontend folder structure, completed static front-end work on the login and register pages
