+++ client/src/pages/Login/Login.css
@@ -65,7 +65,7 @@
 
 .Login p{
     position: relative;
-    font-size: 1.6vw;
+    font-size: 1.4vw;
     align-self: center;
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
     margin-top: 2%;
