+++ client/src/pages/Register/Register.css
@@ -6,9 +6,9 @@
 
 html, body {
     height: 100%;
-    min-width: 50rem; /* Set your desired minimum width */
-    overflow-x: auto; /* Prevent horizontal scrolling */
-    overflow-y: hidden; /* Allow vertical scrolling if content overflows */
+    min-width: 50rem; 
+    overflow-x: auto; 
+    overflow-y: hidden; 
 }
 
 .register-page {
@@ -69,5 +69,5 @@ html, body {
     font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
     font-size: 1.6vw;
     margin-top: 3.5vh;
-    font-size: 1.2vw;
+    font-size: 1.4vw;
 }
